+++ R/rxUiGet.R
@@ -279,6 +279,83 @@ attr(rxUiGet.thetaUpper, "desc") -> "thetaUpper"
 
 #' @export
 #' @rdname rxUiGet
+rxUiGet.lhsVar <- function(x, ...) {
+  .x <- x[[1]]
+  .eta <- get("etaLhsDf", .x)
+  .theta <- get("thetaLhsDf", .x)
+  .cov <- get("covLhsDf", .x)
+  setNames(c(.eta$eta, .theta$theta, .cov$cov),
+           c(.eta$lhs, .theta$lhs, .cov$lhs))
+}
+
+#' @export
+#' @rdname rxUiGet
+rxUiGet.varLhs <- function(x, ...) {
+  .x <- x[[1]]
+  .eta <- get("etaLhsDf", .x)
+  .theta <- get("thetaLhsDf", .x)
+  .cov <- get("covLhsDf", .x)
+  setNames(c(.eta$lhs, .theta$lhs, .cov$lhs),
+           c(.eta$eta, .theta$theta, .cov$cov))
+}
+attr(rxUiGet.varLhs, "desc") <- "var->lhs translation"
+
+#' @export
+#' @rdname rxUiGet
+rxUiGet.lhsEta <- function(x, ...) {
+  .x <- x[[1]]
+  .eta <- get("etaLhsDf", .x)
+  setNames(.eta$eta,.eta$lhs)
+}
+attr(rxUiGet.lhsEta, "desc") <- "lhs->eta translation"
+
+#' @export
+#' @rdname rxUiGet
+rxUiGet.lhsTheta <- function(x, ...) {
+  .x <- x[[1]]
+  .eta <- get("thetaLhsDf", .x)
+  setNames(.eta$theta, .eta$lhs)
+}
+attr(rxUiGet.lhsTheta, "desc") <- "lhs->theta translation"
+
+#' @export
+#' @rdname rxUiGet
+rxUiGet.lhsCov <- function(x, ...) {
+  .x <- x[[1]]
+  .cov <- get("covLhsDf", .x)
+  setNames(.cov$cov, .cov$lhs)
+}
+attr(rxUiGet.lhsCov, "desc") <- "lhs->cov translation"
+
+#' @export
+#' @rdname rxUiGet
+rxUiGet.etaLhs <- function(x, ...) {
+  .x <- x[[1]]
+  .eta <- get("etaLhsDf", .x)
+  setNames(.eta$lhs, .eta$eta)
+}
+attr(rxUiGet.etaLhs, "desc") <- "eta->lhs translation"
+
+#' @export
+#' @rdname rxUiGet
+rxUiGet.thetaLhs <- function(x, ...) {
+  .x <- x[[1]]
+  .theta <- get("thetaLhsDf", .x)
+  setNames(.theta$lhs, .theta$theta)
+}
+attr(rxUiGet.thetaLhs, "desc") <- "theta->lhs translation"
+
+#' @export
+#' @rdname rxUiGet
+rxUiGet.covLhs <- function(x, ...) {
+  .x <- x[[1]]
+  .cov <- get("covLhsDf", .x)
+  setNames(.cov$lhs, .cov$cov)
+}
+attr(rxUiGet.covLhs, "desc") <- "cov->lhs translation"
+
+#' @export
+#' @rdname rxUiGet
 rxUiGet.default <- function(x, ...) {
   .arg <- class(x)[1]
   if (!exists(.arg, envir=x[[1]])) return(NULL)